+++ rhocall/run_annotate_bcfroh.py
@@ -5,8 +5,12 @@ from rhocall import __version__
 logger = logging.getLogger(__name__)
 
 
-def run_annotate_rg(proband_vcf, bcfroh, quality_threshold, flag_upd_at_fraction, output):
-    """Markup VCF file using rho-call BED file."""
+def run_annotate_rg(proband_vcf, bcfroh, quality_threshold, flag_upd_at_fraction, output, select_sample):
+    """Markup VCF file using rho-call BED file.
+
+    If a select_sample is given, only ROH entries for that sample is processed. Otherwise the entries for the
+    first sample encountered is used, even if more are present in the bcftools ROH RG file.
+    """
 
     az_info_header = {
         "ID": "AZ",
@@ -88,7 +92,12 @@ def run_annotate_rg(proband_vcf, bcfroh, quality_threshold, flag_upd_at_fraction
 
         # trust RG entries
         if col[0] == "RG":
+
             sample = str(col[1])
+            if select_sample and sample != select_sample:
+                continue
+            if not select_sample:
+                select_sample = sample
             chrom = str(col[2])
             start = int(col[3])
             end = int(col[4])
@@ -146,11 +155,11 @@ def run_annotate_rg(proband_vcf, bcfroh, quality_threshold, flag_upd_at_fraction
                     # or window is on new chr, and we need to draw new vars to
                     # get there - essentially "before next win"
                     output.write(str(var))
-                    var = next(proband_vcf, False)
                     logger.debug(
                         "Win chr %s not same as var chr %s: keep drawing new vars (end %s)."
                         % (chrom, var.CHROM, var.end)
                     )
+                    var = next(proband_vcf, False)
 
             else:
                 # not found, but passed the due position?!